+++ compiler-core/src/language_server/tests/action.rs
@@ -54,6 +54,7 @@ fn engine_response(src: &str, line: u32) -> engine::Response<Option<Vec<lsp_type
 
 const REMOVE_UNUSED_IMPORTS_TITLE: &str = "Remove unused imports";
 const REMOVE_REDUNDANT_TUPLES: &str = "Remove redundant tuples";
+const ASSIGN_UNUSED_RESULT: &str = "Assign unused Result value to `_`";
 
 fn apply_first_code_action_with_title(src: &str, line: u32, title: &str) -> String {
     let response = engine_response(src, line)
@@ -91,13 +92,13 @@ fn apply_code_edit(
             panic!("Unknown url {}", change_url)
         }
         for edit in change {
-            let start =
-                line_numbers.byte_index(edit.range.start.line, edit.range.start.character) - offset;
-            let end =
-                line_numbers.byte_index(edit.range.end.line, edit.range.end.character) - offset;
+            let start = line_numbers.byte_index(edit.range.start.line, edit.range.start.character)
+                as i32
+                - offset;
+            let end = line_numbers.byte_index(edit.range.end.line, edit.range.end.character) as i32
+                - offset;
             let range = (start as usize)..(end as usize);
-            offset += end - start;
-            offset -= edit.new_text.len() as u32;
+            offset += end - start - edit.new_text.len() as i32;
             result.replace_range(range, &edit.new_text);
         }
     }
@@ -175,6 +176,85 @@ pub fn main() {
         apply_first_code_action_with_title(code, 2, REMOVE_UNUSED_IMPORTS_TITLE),
         expected.replace("%SPACE%", " ")
     )
+}
+
+#[test]
+fn test_assign_unused_result() {
+    let code = "
+pub fn main() {
+    Ok(0)
+    Nil
+}
+";
+
+    insta::assert_snapshot!(apply_first_code_action_with_title(
+        code,
+        2,
+        ASSIGN_UNUSED_RESULT
+    ));
+}
+
+#[test]
+fn test_assign_unused_result_in_block() {
+    let code = "
+pub fn main() {
+    {
+        Ok(0)
+        Nil
+    }
+    Nil
+}
+";
+
+    insta::assert_snapshot!(apply_first_code_action_with_title(
+        code,
+        3,
+        ASSIGN_UNUSED_RESULT
+    ));
+}
+
+#[test]
+fn test_assign_unused_result_only_first_action() {
+    let code = "
+pub fn main() {
+    Ok(0)
+    Ok(0)
+    Nil
+}
+";
+
+    insta::assert_snapshot!(apply_first_code_action_with_title(
+        code,
+        2,
+        ASSIGN_UNUSED_RESULT
+    ));
+}
+
+#[test]
+#[should_panic(expected = "No code action produced by the engine")]
+fn test_assign_unused_result_not_on_return_value() {
+    let code = "
+pub fn main() {
+    Ok(0)
+}
+";
+
+    let _ = apply_first_code_action_with_title(code, 2, ASSIGN_UNUSED_RESULT);
+}
+
+#[test]
+#[should_panic(expected = "No code action produced by the engine")]
+fn test_assign_unused_result_not_on_return_value_in_block() {
+    let code = "
+pub fn main() {
+    let _ = {
+        Ok(0)
+    }
+    Nil
+}
+";
+
+    let _ = apply_first_code_action_with_title(code, 3, ASSIGN_UNUSED_RESULT);
 }
 
 #[test]